requirements.txt
tensorflow==2.16.1
keras==3.3.0
numpy
opencv_python
opencv_python_headless
Pillow
torch==2.3.1
streamlit==1.39.0
ultralytics==8.2.58
transformers
torchvision